Change the main command to be a service-like wrapper
Not an actual service since it needs to run as a user who has a
desktop, which I've no idea how to make happen. This simplifies
controlling it from puppet and should make it easier to track
down when it is/not running, as well as restarting it when needed.

This also changes the logging in the kiosk script to output to
stdout since we're now redirecting that somewhere useful.

Change-Id: Id655bda86f46869019c2438d763264a041f20feb