Add deployment instructions
The README currently provides information about the working setup of
srcomp-kiosk, but does not explain how to deploy it to the Raspberry Pis.
This change adds a brief description of the first-time installation
procedure as well as an update procedure.

The procedures added in this commit have been successfully tested on 6
of the Pis as part of the reimaging process.

Change-Id: I0c80f2fbd40843cfe442ab5f12d2fe8ca640f1e0