Add description of how to reimage the Pis to the readme
It is beneficial to have all of the Pis running consistent setups,
since this allows us to fix problems efficiently without worrying
about differences between individual Pis.
This patch adds a description of the process that was used to
reimage the 6 Pis stored in Southampton, with the hope that this
could become a standard procedure for reimaging the Pis.

Change-Id: I820d85908160f1b4c895325a6f8363f8702a0635